refactor(lib): extract prefs + storage-migrations from storage.js (#826 PR2)
