+++ src/lib/storage.js
@@ -4,19 +4,23 @@
  * Two buckets:
  *   chrome.storage.sync:  user preferences (synced across devices, 100KB quota)
  *   chrome.storage.local: stats and ephemeral state (device-only, 10MB quota)
+ *
+ * #826 PR2: PREF_DEFAULTS + pref accessors live in ./prefs.js.
+ *           One-time migrations live in ./storage-migrations.js.
+ *           Full pre-split public API is re-exported here so all 11 importers
+ *           keep working unchanged.
  */
 
-// Per-device consent overlay for getPrefs (#355, ADR-0001).
-import { getConsent } from "./consent-storage.js";
-// Per-device pref overrides on top of synced behavioural prefs (#364).
-import { getOverrides as getPerDeviceOverrides } from "./per-device-prefs.js";
-// Consent version-comparison for feature gating during hard re-onboard (#370).
-import { evaluate as evaluateConsentPolicy } from "./consent-policy.js";
-// E2E fixture overrides (#407). Returns null in production.
-import { getTestFixtures } from "./test-fixtures.js";
 // Allowlist guard for the per-shortener counters (ADR-0004 phase 4, #700):
 // only the eight GENERIC_SHORTENERS may be recorded — never arbitrary hosts.
 import { isGenericShortener } from "./opaque-networks.js";
+
+// ── Re-exports: prefs domain ──────────────────────────────────────────────────
+// Keeps the full pre-split public API available at the storage.js import path.
+export { PREF_DEFAULTS, getPrefs, setPrefs } from "./prefs.js";
+
+// ── Re-exports: one-time migrations ──────────────────────────────────────────
+export { migrateStatsToLocal, migrateLegacyProxyPref } from "./storage-migrations.js";
 
 // ── Firefox MV2 compat: chrome.* APIs must return Promises ──────────────────
 //
@@ -81,184 +85,6 @@ import { isGenericShortener } from "./opaque-networks.js";
     wrapMethod(chrome.contextMenus, "removeAll");
   }
 })();
-
-// ── Sync: user preferences ──────────────────────────────────────────────────
-
-export const PREF_DEFAULTS = {
-  enabled: true,
-  injectOwnAffiliate: false,  // set to true only if user opts in during onboarding (#224)
-  notifyForeignAffiliate: false,
-  stripAllAffiliates: false,
-  blacklist: [],     // e.g. ["amazon.es", "booking.com::aid::123456"]
-  whitelist: [],     // e.g. ["amazon.es::tag::youtuber-21"]
-  customParams: [],  // e.g. ["ref_code", "promo_id"]
-  dnrEnabled: true,
-  contextMenuEnabled: true,
-  blockPings: true,
-  ampRedirect: true,
-  unwrapRedirects: true,
-  language: "en",
-  onboardingDone: false,
-  consentVersion: null,   // e.g. "1.0". Bump to re-trigger onboarding on ToS changes.
-  consentDate: null,      // Unix timestamp (ms) of when the user accepted
-  disabledCategories: [],  // e.g. ["utm", "ads"]. Params in these categories are not stripped.
-  toastDuration: 15,  // seconds: how long the affiliate notification stays visible
-  paramBreakdown: true,
-  showReportButton: true,
-  domainStats: true,
-  // Remote rules toggle — lives in sync so the opt-in preference follows the user
-  // across devices. Default false per REQ-OPT-1: zero network activity on fresh install.
-  remoteRulesEnabled: false,
-  // Honor Creator Mode toggle (#435, B12). Pure plumbing for B13/B14: no
-  // behaviour change. Default false so existing users see no functional
-  // difference. The feature is opt-in because honoring creator referral
-  // chains may route through redirect networks the user did not consent to
-  // contact otherwise.
-  honorCreatorMode: false,
-  // Per-creator allowlist (#445, B13). Referrer-domain-shaped strings
-  // (e.g., "youtube.com/@LinusTechTips", "dot-css-news.com") that the user
-  // has explicitly opted into for Honor Creator Mode. Empty by default.
-  // Capped at 100 entries (storage hygiene): 100 × ~80 chars ≈ 8 KB, within
-  // sync's 8 KB-per-item / 100 KB-total budget. CRUD lives in
-  // src/lib/creator-allowlist.js (pure module, immutable arrays).
-  creatorAllowlist: [],
-  // Canonical Extractor toggle (#442, B7). Default ON: when the wrapper
-  // engine detects an opaque wrapper (host matched but no destination in
-  // the URL), the cleaner consults a content-script-supplied "canonical
-  // bundle" (<link rel=canonical> + JSON-LD @id) BEFORE giving up. Disable
-  // here to bypass that tier entirely without uninstalling content scripts.
-  canonicalExtractorEnabled: true,
-  // Cross-Site Frequency Tracker toggle (#446, B16). Default ON: a local-
-  // only correlation map of (paramName, hashedValue) per first-party
-  // domain, used to surface likely cross-site identifiers in the popup.
-  // Privacy-sensitive enough to deserve its own toggle even though the
-  // data never leaves the device — turning it off makes observe() a
-  // no-op and hides the freq subgroup in the suspicious-params section.
-  crossSiteFrequencyEnabled: true,
-  // Attribution Ledger persistence (#460, A2). When ON (default), the SW
-  // writes the rolling ring buffer of recent navigation events to
-  // chrome.storage.local under "attributionLedger" so the popup can
-  // render a "Recent activity" section that survives SW restarts. When
-  // OFF, the writer is gated to a no-op; the section just stays empty.
-  // Privacy-sensitive (it carries URLs), so we expose it as its own
-  // toggle even though the data is local-only.
-  attributionLedgerEnabled: true,
-  // EXPERIMENTAL shape-based param heuristic (#544). Default OFF: a multi-
-  // signal heuristic that strips params whose VALUE SHAPE matches a tracker
-  // pattern (suspicious key prefix + length>16 + Shannon entropy>4.0 +
-  // base64/hex/uuid charset — ALL four required). False-positive risk is
-  // real (auth tokens / session IDs LOOK like trackers), so it ships behind
-  // a flag and is gated by a hard-coded allowlist of well-known oauth /
-  // session keys (state, code, csrf_token, access_token, …) that never
-  // strip even when all four signals fire. With the flag OFF, behaviour is
-  // byte-identical to the #530 baseline (the benchmark stays 117/117).
-  experimentalParamClassesEnabled: false,
-  // User-promoted custom strip rules (#536). Populated by the popup's
-  // "Strip locally" button on flagged Suspicious-params rows. Each entry
-  // is a bare param name (lowercased on read by the cleaner) that
-  // processUrl strips on EVERY host — the user has explicitly trusted
-  // the rule. Affiliate-preservation still wins (the affiliateParamSet
-  // skip in stripTrackingParams runs before custom-rule matching), so a
-  // user can never accidentally strip their own creator's referral tag.
-  // Lives in sync so the rule list follows the user across devices.
-  // Default empty — opt-in by user click only.
-  userCustomRules: [],
-  // Follow shortener redirects natively (ADR-0004 phase 2, #699; renamed from
-  // privacyProxyEnabled in phase 5, #701). When ON, MUGA resolves the eight
-  // generic shorteners in-browser via fetch(redirect:"manual"). Default OFF:
-  // requires the eight shortener host permissions, granted from the options toggle.
-  // Migration: on startup, if chrome.storage.sync contains privacyProxyEnabled=true,
-  // this field is set to true and the old key is deleted (see migrateLegacyProxyPref).
-  followShortenersEnabled: false,
-  // NOTE (ADR-0004 phase 5, 2026-06-01): privacyProxyEnabled was the Privacy Proxy
-  // toggle removed in phase 5. Retained as a deprecation comment only — do NOT add
-  // it back to PREF_DEFAULTS. Any live value is migrated to followShortenersEnabled
-  // on first startup by migrateLegacyProxyPref().
-  //
-  // NOTE (ADR-0004 phase 5, 2026-06-01): useNativeShortenerResolution was the
-  // dual-path selector removed in phase 5. Native resolution is now the ONLY path.
-  // Do NOT add it back to PREF_DEFAULTS.
-};
-
-/**
- * Reads all user preferences. Behavioural prefs come from
- * `chrome.storage.sync` (cross-device). Consent fields
- * (`onboardingDone`, `consentVersion`, `consentDate`) are overlaid
- * from per-device `chrome.storage.local` via consent-storage (#355,
- * ADR-0001). Per-device behavioural overrides (`injectOwnAffiliate`,
- * `remoteRulesEnabled` after a user declines a sync-inherited prompt)
- * are overlaid from per-device-prefs (#364) — local wins.
- *
- * Reads in parallel for minimum latency.
- *
- * @returns {Promise<object>} Preferences merged with PREF_DEFAULTS.
- */
-export async function getPrefs() {
-  try {
-    const [sync, consent, overrides, fixtures] = await Promise.all([
-      new Promise((resolve, reject) => {
-        chrome.storage.sync.get(PREF_DEFAULTS, (result) => {
-          if (chrome.runtime.lastError) reject(chrome.runtime.lastError);
-          else resolve(result);
-        });
-      }),
-      getConsent(),
-      getPerDeviceOverrides(),
-      getTestFixtures(),
-    ]);
-
-    // Consent overlay (#355). Local wins over sync.
-    const overlay = {};
-    if (consent.onboardingDone) overlay.onboardingDone = true;
-    if (consent.consentVersion !== null) overlay.consentVersion = consent.consentVersion;
-    if (consent.consentDate !== null) overlay.consentDate = consent.consentDate;
-
-    // Hard-reonboard gate (#370). When ConsentPolicy says material change
-    // pending, force `onboardingDone: false` so existing feature gates
-    // (`if (!prefs.onboardingDone) return`) bail until the user re-accepts.
-    // Soft re-onboard does NOT gate features — the user's prior consent
-    // remains valid for previously accepted behaviour.
-    // Under e2e fixtures (#407), the gate fires against the fixture
-    // manifest + required version so tests can drive the dormant path.
-    const policy = evaluateConsentPolicy({
-      stored: consent,
-      ...(fixtures?.requiredConsentVersion ? { requiredVersion: fixtures.requiredConsentVersion } : {}),
-      ...(fixtures?.consentManifest ? { manifest: fixtures.consentManifest } : {}),
-    });
-    if (policy.status === "hard-reonboard") {
-      overlay.onboardingDone = false;
-    }
-
-    // Per-device pref overlay (#364). Any key set in overrides wins
-    // over sync. Boolean shape is enforced at the source (overrides
-    // can only be set via per-device-prefs.setOverrides).
-    return { ...sync, ...overlay, ...overrides };
-  } catch (err) {
-    console.error("[MUGA] getPrefs failed:", err);
-    return { ...PREF_DEFAULTS };
-  }
-}
-
-/**
- * Writes a partial preferences object to chrome.storage.sync.
- * @param {object} partial - Key/value pairs to merge into stored prefs.
- * @returns {Promise<void>}
- */
-export async function setPrefs(partial) {
-  try {
-    return await new Promise((resolve, reject) => {
-      chrome.storage.sync.set(partial, () => {
-        if (chrome.runtime.lastError) {
-          reject(chrome.runtime.lastError);
-        } else {
-          resolve();
-        }
-      });
-    });
-  } catch (err) {
-    console.error("[MUGA] setPrefs failed:", err);
-  }
-}
 
 // ── Local: stats and nudge state ─────────────────────────────────────────────
 
@@ -764,94 +590,5 @@ export function incrementShortenerStat(host, outcome) {
   else _pendingShortenerStats[host].fail += 1;
   if (!_shortenerStatsFlushTimer) {
     _shortenerStatsFlushTimer = setTimeout(_flushShortenerStats, 50);
-  }
-}
-
-// ── One-time migration ────────────────────────────────────────────────────────
-
-/**
- * One-time migration: moves stats out of chrome.storage.sync into
- * chrome.storage.local. Safe to call on every startup. Exits immediately
- * if migration already done or no old data exists.
- */
-export async function migrateStatsToLocal() {
-  const syncData = await new Promise((resolve, reject) =>
-    chrome.storage.sync.get({ stats: null, firstUsed: null, nudgeDismissed: null }, (result) => {
-      if (chrome.runtime.lastError) reject(chrome.runtime.lastError);
-      else resolve(result);
-    })
-  ).catch(() => ({ stats: null, firstUsed: null, nudgeDismissed: null }));
-
-  const hasOldStats =
-    syncData.stats !== null ||
-    syncData.firstUsed !== null ||
-    syncData.nudgeDismissed !== null;
-
-  if (!hasOldStats) return;
-
-  // Copy to local (only if local doesn't already have data)
-  const localData = await getStats();
-  const merged = {
-    stats: syncData.stats ?? localData.stats,
-    firstUsed: syncData.firstUsed ?? localData.firstUsed,
-    nudgeDismissed: syncData.nudgeDismissed ?? localData.nudgeDismissed,
-  };
-  await setStats(merged);
-
-  // Remove from sync
-  await new Promise((resolve, reject) =>
-    chrome.storage.sync.remove(["stats", "firstUsed", "nudgeDismissed"], () => {
-      if (chrome.runtime.lastError) reject(chrome.runtime.lastError);
-      else resolve();
-    })
-  ).catch(() => {}); // best-effort: old keys already migrated, removal is non-critical
-}
-
-/**
- * One-time migration (ADR-0004 phase 5, #701): renames `privacyProxyEnabled` →
- * `followShortenersEnabled` in chrome.storage.sync. Safe to call on every
- * startup. Exits immediately if the old key is absent.
- *
- * If the user had `privacyProxyEnabled = true` they were using native shortener
- * resolution (the default since phase 4 / 2.2.0-beta.1), so we preserve their
- * intent by setting `followShortenersEnabled = true`. A false value needs no
- * migration because `followShortenersEnabled` already defaults to false.
- */
-export async function migrateLegacyProxyPref() {
-  try {
-    const data = await new Promise((resolve, reject) =>
-      chrome.storage.sync.get({ privacyProxyEnabled: null }, (result) => {
-        if (chrome.runtime.lastError) reject(chrome.runtime.lastError);
-        else resolve(result);
-      })
-    ).catch(() => ({ privacyProxyEnabled: null }));
-
-    if (data.privacyProxyEnabled === null) return; // key absent — nothing to do
-
-    const updates = {};
-    if (data.privacyProxyEnabled === true) {
-      // Preserve user's intent: they had the feature enabled.
-      updates.followShortenersEnabled = true;
-    }
-    // Write followShortenersEnabled only when migrating a `true` value; a false
-    // or absent old value needs no write (the new key already defaults to false).
-    // The old key is removed unconditionally below regardless.
-    if (Object.keys(updates).length > 0) {
-      await new Promise((resolve, reject) =>
-        chrome.storage.sync.set(updates, () => {
-          if (chrome.runtime.lastError) reject(chrome.runtime.lastError);
-          else resolve();
-        })
-      );
-    }
-    // Remove the deprecated key.
-    await new Promise((resolve) =>
-      chrome.storage.sync.remove("privacyProxyEnabled", () => {
-        void chrome.runtime.lastError; // non-critical
-        resolve();
-      })
-    );
-  } catch {
-    // Migration is best-effort — a failure here must never break startup.
   }
 }
